Add files via upload
This is an application used to manage a scooter rental business. The functions are divided according to the user's roles: manager or customer. The administrator has the possibility to add, delete or modify scooters, to filter or sort them, but also to search for them according to location. The customer can search, filter, sort and rent a scooter.

Also, the user has the possibility to choose the way in which the data is saved. It is stored either volatile (the information is deleted at the end of the program execution) or persistent (the data is saved in a csv file).